+++ gm-utils.js
@@ -880,6 +880,7 @@
         notify_shadowfront_reminder_5: 'true',
         notify_shadowfront_start: 'true',
         notify_calamity_10: 'true',
+        notify_gvg_daily_tasks: 'true',
         notify_gvg_pvp: 'true',
         notify_svs_garrison: 'true',
         notify_svs_pvp: 'true',